moved inverter and ac-filter classes from common.model._model into separate modules
