+++ motulator/common/model/__init__.py
@@ -1,8 +1,9 @@
 """Common functions and classes for continuous-time system models."""
 from motulator.common.model._simulation import Simulation
 from motulator.common.model._model import (
-    Delay, CarrierComparison, zoh, Model, Subsystem,
-    Inverter, DiodeBridge, ACFilter)
+    Delay, CarrierComparison, zoh, Model, Subsystem)
+from motulator.common.model._converter import Inverter, DiodeBridge
+from motulator.common.model._ac_filter import ACFilter
 
 __all__ = [
     "Simulation",
